+++ src/app/docs/components/button/border-glow-button/page.mdx
@@ -1,0 +1,8 @@
+export const metadata = {
+  title: 'Border Glow Button',
+  description: 'An animated button featuring a mesmerizing border effect that glows around your mouse cursor and adds a unique visual flair to your UI. Built using Tailwind CSS & Framer Motion for React and libraries like Next.js'
+}
+
+<BackButton href="/docs/components/button" />
+
+<ComponentPreview path="components/button/BorderGlowButton" usingFramer/>
