--- src/app/docs/effects/border-glow/page.mdx
@@ -1,22 +1,0 @@
-
-import BorderGlow from '@/showcase/effects/BorderGlow'
-
-export const metadata = {
-  title: 'Border Glow',
-  description: 'Adds a border that glows around an element based on the mouse position',
-}
-
-# Border Glow Effect
-
-<div className="mt-4">Adds a gradual transparency to images</div>
-
-
-<ComponentPreview
- preview={
-    <div className="flex flex-col gap-2">
-      <BorderGlow>Border Glow</BorderGlow>
-      <BorderGlow>Border Glow</BorderGlow>
-    </div>
-  }
-  path="effects/BorderGlow"
-/>
