+++ src/showcase/ui-group/EffectCards.tsx
@@ -1,7 +1,6 @@
 import Image from 'next/image'
 import Link from 'next/link'
 import React from 'react'
-import BorderGlow from '@/showcase/effects/BorderGlow'
 
 const data = [
   {
@@ -16,17 +15,6 @@ const data = [
     image: '/images/ui/gradients.png',
     link: '/docs/effects/gradients',
   },
-  {
-    id: 3,
-    title: 'Border Glow',
-    preview: (
-      <div className="flex flex-col gap-2">
-        <BorderGlow>Border Glow</BorderGlow>
-        <BorderGlow>Border Glow</BorderGlow>
-      </div>
-    ),
-    link: '/docs/effects/border-glow',
-  },
 ]
 
 const EffectCards = () => {
@@ -37,16 +25,13 @@ const EffectCards = () => {
           <Link href={item.link} key={item.id}>
             <div className="overflow group rounded-xl border border-white ring-1 ring-zinc-200 transition-all ease-in-out hover:cursor-pointer">
               <div className="flex h-[9rem] items-center justify-center rounded-t-xl border-b bg-gray-50 text-xs text-gray-400 transition-all ease-in-out group-hover:bg-gray-100 md:h-[12rem]">
-                {item.preview ? item.preview : null}
-                {item.image ? (
-                  <Image
-                    width={200}
-                    height={200}
-                    src={item.image}
-                    alt={item.title}
-                    className="h-auto w-[150px] duration-300 ease-in-out group-hover:scale-110"
-                  />
-                ) : null}
+                <Image
+                  width={200}
+                  height={200}
+                  src={item.image}
+                  alt={item.title}
+                  className="h-auto w-[150px] duration-300 ease-in-out group-hover:scale-110"
+                />
               </div>
               <div className="w-full p-4 text-sm font-medium text-gray-800">
                 {item.title}
